backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0

# Validation
pydantic>=2.7.0

# Auth & crypto
cryptography>=42.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0

# Scheduler  ← NEW in Sprint 1
apscheduler>=3.10.0

# AI (optional — chat router falls back to rule-based without it)
anthropic>=0.28.0

# Connectors
pyodbc>=5.0.0
pymssql>=2.2.0
boto3>=1.34.0
azure-storage-blob>=12.19.0
paramiko>=3.4.0
requests>=2.31.0
urllib3>=2.2.0

# Export
openpyxl>=3.1.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-html>=4.0.0

# MCP — standalone MCP servers + agentic tool-use in chat router
# Install: pip install mcp
mcp>=1.0.0
